samples/rag-ai-agent-dataverse/agent/requirements.txt
requests
python-dotenv
azure-identity
openai
